Namespace PresenterSpecHelper to Spec::Support:PresenterHelper
